Require at least Node 14.
The current code uses dynamic import, which only works adequately well on
Node 12.17 and newer. The CircleCI images for specific minor versions of
Node don't play well with the browser-tools orb. Given the newness of
jasmine-browser-runner and its tiny userbase, it seems better to draw the
line at 14.0 than to solve that problem.

[#178075437]