Support loading custom reporters from ES modules
Reporter modules are now loaded via dynamic import() instead of require(),
so they can be either ES modules or CommonJS. As before, the module's
default export is expected to be a constructor that can be called without
arguments and returns a reporter.

This is a breaking change. Reporter names specified in config files and via
the --reporter CLI arg now need to conform to the rules for ES module import
specifiers: <https://nodejs.org/api/esm.html#esm_import_specifiers>. This
applies even if the reporter is a CommonJS module. In practice this means
that some users may need to append '.js' to their reporter names.

[#178599159]