Opt-in support for top-level-await
This can be enabled by adding enableTopLevelAwait: true to the configuration.
It's off by default because HTTP requests for scripts are performed
sequentially rather than potentially in parallel. This comes at a cost in
performance but is the only way to ensure a stable order when randomization
is disabled or a random seed is specified.